launchpad/requirements/core.txt
Django==3.2.11
Celery==5.2.3
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2==2.8.6
django-redis==5.2.0
django-celery-beat==2.2.1
kubernetes==21.7.0
django-celery-results==2.2.0
drf-yasg==1.20.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
